first offline version

+++ app/views.py
@@ -95,7 +95,7 @@ def RemoveOldToken():
                     os.remove(file_location)
 
 
-@login_required(login_url='app:login')                
+@login_required(login_url='app:login')
 def SendEmailHotmail(email,subject,message):
     ol=win32com.client.Dispatch("outlook.application",pythoncom.CoInitialize())
     olmailitem=0x0 
@@ -304,7 +304,6 @@ def Register(request):
         return render(request, 'registration/result.html', {'result': 'denied'})
 
 
-@login_required(login_url='app:login')
 def Login(request):
 
     if request.method == 'POST':
@@ -368,6 +367,7 @@ def generalInsert(request, mainField, baseDic, model, addModel, savePoint):
     for i in range(len(request.POST.getlist(mainField))):
         dic = {'csrfmiddlewaretoken': request.POST['csrfmiddlewaretoken']}
         dic.update(baseDic)
+        dic.update({'isOffline': True})
         for field in model._meta.local_fields:
             if field.name in request.POST:
                 dic[field.name] = request.POST.getlist(field.name)[i]
@@ -1186,6 +1186,7 @@ def generalDelete(modelName, objectId):
     deletedObject= DeletedObjects()
     deletedObject.modelName= modelName
     deletedObject.objectId = objectId
+    deletedObject.isOffline= True
     deletedObject.save()
 
 
@@ -1691,7 +1692,7 @@ def pullData(request):
                     for model in apps.get_models():
                         if not model.__name__ in ['LogEntry', 'Permission', 'Group', 'User', 'ContentType', 'Session', 'LastPull', 'DeletedObjects', 'UploadedFile']:
                             serializerClass = getSerializer(model.__name__)
-                            added = serializerClass(model.objects.filter(createdDate__gte=lastPullDate), many= True).data
+                            added = serializerClass(model.objects.filter(isOffline=True), many= True).data
                             updated =serializerClass(model.objects.filter(Q(lastModifiedDate__gte=lastPullDate) & ~Q(createdDate__gte=lastPullDate) ), many= True).data
                             deleted = SerializerDeletedObjects( DeletedObjects.objects.filter(modelName=model.__name__, createdDate__gte=lastPullDate), many= True).data
                             data.update( {model.__name__: {'updated':updated, 'added':added, 'deleted': deleted} })
@@ -1706,8 +1707,9 @@ def pullData(request):
                     transaction.savepoint_rollback(savePoint)
                     return render(request, 'registration/result.html', {'result': 'done'}) 
             #  temp = LastPull.objects.filter(pk=1).update(lastPullDate=datetime.datetime.now)
-             temp = LastPull.objects.get(pk=1)
+             temp = LastPull.objects.get(userId_id__is_superuser=1)
              temp.lastPullDate=datetime.datetime.now
+             temp.waitingMerge = True
              LastPull.save(self=temp)
              return render(request, 'registration/result.html', {'result': 'done'})
            
@@ -1722,21 +1724,68 @@ def pullData(request):
 @login_required(login_url='app:login')
 def generalPushAdd(request ,added, addModel, modelName, idMap, savePoint):
     id = None
-    oldId = added['id']
-    del added['id']
+    haveId = 'id' in added
+    oldId= None
+    if haveId:
+        oldId = added['id']
+        del added['id']
+    print(modelName)
     dic = {'csrfmiddlewaretoken': get_token(request)}
     dic.update(added)
     dic.update({'isOffline': False})
     form = addModel(dic)
+    print(1)
     if form.is_valid():
-        id = form.save()
-        if not idMap[modelName]:
-            idMap[modelName] = {}
-        idMap[modelName].update({oldId: id})
+        id = form.save(commit=False)
+        if haveId:
+            id.pk = oldId
+        id.save()
+        print(2)
+        if haveId:
+            if not modelName in idMap:
+                idMap[modelName] = {}
+            idMap[modelName].update({oldId: id})
+        
     else:
         transaction.savepoint_rollback(savePoint)
         return form.errors
     return id
+
+
+
+def generalPushAddHub(request, added, addModel, modelName, idMap, savePoint):
+    #Demonstrator
+    print(modelName)
+
+    if modelName in ['Dispatch', 'GraduateStudies', 'CertificateOfExcellence', 'AdjectiveChange']:
+        #studentId
+        if 'Demonstrator' in idMap:
+            if added['studentId'] in idMap['Demonstrator']:
+                added['studentId'] = idMap['Demonstrator'][added['studentId']]
+    elif modelName == 'Nomination':
+        #nominationDecision
+        if 'Demonstrator' in idMap:
+            if added['nominationDecision'] in idMap['Demonstrator']:
+                added['nominationDecision'] = idMap['Demonstrator'][added['nominationDecision']]
+    elif modelName == 'UniversityDegree':
+        #universityDegree
+        if 'Demonstrator' in idMap:
+            if added['universityDegree'] in idMap['Demonstrator']:
+                added['universityDegree'] = idMap['Demonstrator'][added['universityDegree']]
+
+    #Dispatch
+    elif modelName in ['Report', 'Extension', 'Freeze', 'DurationChange', 'AlimonyChange', 'UniversityChange', 'SpecializationChange']:
+        #dispatchDecisionId
+        if 'Dispatch' in idMap:
+            if added['dispatchDecisionId'] in idMap['Dispatch']:
+                added['dispatchDecisionId'] = idMap['Dispatch'][added['dispatchDecisionId']]
+    elif modelName == 'Regularization':
+        #regularizationDecisionId
+        if 'Dispatch' in idMap:
+            if added['regularizationDecisionId'] in idMap['Dispatch']:
+                added['regularizationDecisionId'] = idMap['Dispatch'][added['regularizationDecisionId']]
+
+    return generalPushAdd(request, added , addModel, modelName, idMap, savePoint)
 
 
 @login_required(login_url='app:login')
@@ -1758,30 +1807,30 @@ def generalUpdateHub(request, added, obj, addModel, modelName, idMap, savePoint)
     #Demonstrator
     if modelName in ['Dispatch', 'GraduateStudies', 'CertificateOfExcellence', 'AdjectiveChange']:
         #studentId
-        if idMap[modelName]:
-            if idMap[modelName][added['studentId']]:
+        if modelName in idMap:
+            if added['studentId'] in idMap[modelName]:
                 added['studentId'] = idMap[modelName][added['studentId']]
     elif modelName == 'Nomination':
         #nominationDecision
-        if idMap[modelName]:
-            if idMap[modelName][added['nominationDecision']]:
+        if modelName in idMap:
+            if added['nominationDecision'] in idMap[modelName]:
                 added['nominationDecision'] = idMap[modelName][added['nominationDecision']]
     elif modelName == 'UniversityDegree':
         #universityDegree
-        if idMap[modelName]:
-            if idMap[modelName][added['universityDegree']]:
+        if modelName in idMap:
+            if added['universityDegree'] in idMap[modelName]:
                 added['universityDegree'] = idMap[modelName][added['universityDegree']]
 
     #Dispatch
     elif modelName in ['Report', 'Extension', 'Freeze', 'DurationChange', 'AlimonyChange', 'UniversityChange', 'SpecializationChange']:
         #dispatchDecisionId
-        if idMap[modelName]:
-            if idMap[modelName][added['dispatchDecisionId']]:
+        if modelName in idMap:
+            if added['dispatchDecisionId'] in idMap[modelName]:
                 added['dispatchDecisionId'] = idMap[modelName][added['dispatchDecisionId']]
     elif modelName == 'Regularization':
         #regularizationDecisionId
-        if idMap[modelName]:
-            if idMap[modelName][added['regularizationDecisionId']]:
+        if modelName in idMap:
+            if added['regularizationDecisionId'] in idMap[modelName]:
                 added['regularizationDecisionId'] = idMap[modelName][added['regularizationDecisionId']]
 
     return generalPushUpdate(request, added, obj, addModel, savePoint)
@@ -1835,12 +1884,23 @@ def pushData(request):
                     idMap = {}
                     with open('uploads/synchronization.json', 'r') as f:
                         data = load(f)
+                    
+                    #delete offline changes before merge
+                    for model in apps.get_models():
+                        if not model.__name__ in ['LogEntry', 'Permission', 'Group', 'User', 'ContentType', 'Session', 'LastPull', 'UploadedFile']:
+                            delObjs= model.objects.filter(isOffline=True).delete()
+
+      
                     for model in apps.get_models():
                         if not model.__name__ in ['LogEntry', 'Permission', 'Group', 'User', 'ContentType', 'Session', 'LastPull', 'DeletedObjects', 'UploadedFile']:
                             addModel= getForm(model.__name__)
                             # add
                             for added in data[model.__name__]['added']:
-                                id = generalPushAdd(request, added , addModel, model.__name__, idMap, savePoint)
+                                if 'id' in added:
+                                    isExist = model.objects.filter(pk=added['id'])
+                                    if isExist.count()>0:
+                                        continue
+                                id = generalPushAddHub(request, added , addModel, model.__name__, idMap, savePoint)
                                 if type(id) == ErrorDict: return render(request, 'registration/result.html', {'result': id})
                                 if model.__name__ in ['Dispatch', 'Freeze', 'Extension', 'DurationChange']:
                                     dispatchId = 1
@@ -1892,8 +1952,7 @@ def pushData(request):
                                         demonstrator.currentAdjective = updated.adjectiveChangeAdjective
                                         Demonstrator.full_clean(self=demonstrator)
                                         Demonstrator.save(self=demonstrator)
-
-                                   
+                 
                     for model in apps.get_models():
                         if not model.__name__ in ['LogEntry', 'Permission', 'Group', 'User', 'ContentType', 'Session', 'LastPull', 'DeletedObjects', 'UploadedFile']:
                             addModel= getForm(model.__name__)
@@ -1902,6 +1961,9 @@ def pushData(request):
                                 if idMap[model.__name__]:
                                     if idMap[model.__name__][deleted.id]:
                                         deleted.id = idMap[model.__name__][deleted.id]
+                                isExist = model.objects.filter(pk=deleted.id)
+                                if isExist.count()==0:
+                                    continue
                                 deletedObj= model.objects.filter(pk=deleted.id).delete()
                                 if model.__name__ in [ 'Freeze', 'Extension', 'DurationChange']:
                                     dispatchId = deletedObj.dispatchDecisionId
@@ -1923,12 +1985,13 @@ def pushData(request):
                                 deletedObject= DeletedObjects()
                                 deletedObject.modelName= model.__name__
                                 deletedObject.objectId = deleted.id
-                                deletedObject.isOffline = False
                                 deletedObject.save()
 
-                                
-
-
+                    print('finish')
+                    temp = LastPull.objects.get(userId_id__is_superuser=1)
+                    temp.waitingMerge = False
+                    LastPull.save(self=temp)
+                    print('final')
                             
                     return render(request, 'registration/result.html', {'result': 'done'})
                 except Exception as e:
